+++ board/project/protocol/jetiex.c
@@ -555,8 +555,8 @@ void jeti_set_config(sensor_jetiex_t **sensor) {
         parameter.protocol = config->gps_protocol;
         parameter.baudrate = config->gps_baudrate;
         parameter.rate = config->gps_rate;
-        parameter.lat = malloc(sizeof(double));
-        parameter.lon = malloc(sizeof(double));
+        parameter.lat = malloc(sizeof(float));
+        parameter.lon = malloc(sizeof(float));
         parameter.alt = malloc(sizeof(float));
         parameter.spd = malloc(sizeof(float));
         parameter.cog = malloc(sizeof(float));
