+++ board/project/protocol/smartport.c
@@ -1538,8 +1538,8 @@ static void set_config(smartport_parameters_t *parameter) {
         parameter.protocol = config->gps_protocol;
         parameter.baudrate = config->gps_baudrate;
         parameter.rate = config->gps_rate;
-        parameter.lat = malloc(sizeof(double));
-        parameter.lon = malloc(sizeof(double));
+        parameter.lat = malloc(sizeof(float));
+        parameter.lon = malloc(sizeof(float));
         parameter.alt = malloc(sizeof(float));
         parameter.spd = malloc(sizeof(float));
         parameter.cog = malloc(sizeof(float));
